Just some trivial changes.

+++ src/buf.h
@@ -11,6 +11,8 @@
 #include <pthread.h>
 #include "defs.h"
 
+// constants {{{
+
 // buffer size
 #define BUF_SAMPLES_LOG 6u
 #define BUF_SAMPLES (1u << BUF_SAMPLES_LOG)
@@ -28,6 +30,10 @@
 #define BUFPOOL_BYTES (1u << BUFPOOL_BYTES_LOG)
 #define BUFPOOL_MASK (BUFPOOL_BYTES - 1)
 
+// }}}
+
+// types {{{
+
 typedef struct bufpool_s* bufpool_t;
 
 typedef float samp_t;
@@ -44,6 +50,10 @@ struct bufpool_s {
     pthread_mutex_t mutex;
     size_t free;
 };
+
+// }}}
+
+// bufpools {{{
 
 extern bufpool_t bufpool_create (void);
 extern void bufpool_destroy (bufpool_t);
@@ -64,6 +74,11 @@ extern buf_t buf_alloc (bufpool_t); // refcount = 1
 extern void buf_acquire (buf_t);
 extern void buf_release (buf_t);
 
+// }}}
+
+// buf operations {{{
+
+// const {{{
 MACRO void
 buf_const (buf_t b, samp_t x, size_t n)
 {
@@ -71,31 +86,36 @@ buf_const (buf_t b, samp_t x, size_t n)
     for (i = 0; i < n; i++)
         b.xs[i] = x;
 }
+// }}}
 
+// lin {{{
 MACRO samp_t
 buf_lin (buf_t b, samp_t x0, samp_t dx, size_t n)
 {
     samp4_t x4 = {x0, x0 + dx, x0 + 2*dx, x0 + 3*dx};
-    //samp4_t dx4 = 4 * (samp4_t) dx;
     size_t i;
     for (i = 0; 4*i < n; i++, x4 += 4 * dx)
         b.x4s[i] = x4;
     return x0 + n * dx;
 }
+// }}}
 
+// xdec {{{
 MACRO samp_t
 buf_xdec (buf_t b, samp_t x0, samp_t xinf, samp_t tau, size_t n)
 {
-    samp_t f1 = exp (-1/tau);
-    samp_t f4 = exp (-4/tau);
+    samp_t f1 = expf (-1/tau);
+    samp_t f4 = expf (-4/tau);
     samp_t h = x0 - xinf;
     samp4_t z4 = {1, f1, f1*f1, f1*f1*f1};
     size_t i;
     for (i = 0; 4*i < n; i++, z4 *= f4)
         b.x4s[i] = xinf + h * z4;
-    return xinf + h * exp (-n / tau);
+    return xinf + h * expf (-n / tau);
 }
+// }}}
 
+// add {{{
 MACRO void
 buf_add (buf_t acc, buf_t b, size_t n)
 {
@@ -103,7 +123,9 @@ buf_add (buf_t acc, buf_t b, size_t n)
     for (i = 0; 4*i < n; i++)
         acc.x4s[i] += b.x4s[i];
 }
+// }}}
 
+// mul {{{
 MACRO void
 buf_mul (buf_t acc, buf_t b, size_t n)
 {
@@ -111,7 +133,9 @@ buf_mul (buf_t acc, buf_t b, size_t n)
     for (i = 0; 4*i < n; i++)
         acc.x4s[i] *= b.x4s[i];
 }
+// }}}
 
+// scale {{{
 MACRO void
 buf_scale (buf_t b, buf_t src, samp_t s, size_t n)
 {
@@ -119,14 +143,19 @@ buf_scale (buf_t b, buf_t src, samp_t s, size_t n)
     for (i = 0; 4*i < n; i++)
         b.x4s[i] = s * src.x4s[i];
 }
+// }}}
 
+// cos2pi {{{
 MACRO void
 buf_cos2pi (buf_t b, buf_t src, size_t n)
 {
     size_t i;
     for (i = 0; i < n; i++)
-        b.xs[i] = cos ((2 * (float) M_PI) * src.xs[i]);
+        b.xs[i] = cosf ((2 * (float) M_PI) * src.xs[i]);
 }
+// }}}
+
+// }}}
 
 #endif /* koar/buf.h */
